new documents

+++ doc/sphinx_util.py
@@ -6,22 +6,21 @@ import docutils
 import subprocess
 
 
+READTHEDOCS_BUILD = (os.environ.get('READTHEDOCS', None) == 'True')
+
 def run_build_mxnet(folder):
     """Run the doxygen make command in the designated folder."""
     try:
-        subprocess.call('cd ..; rm -rf dmlc-core;' +
-                        'git clone https://github.com/dmlc/dmlc-core', shell = True)
-        subprocess.call('cd ..; rm -rf mshadow;' +
-                        'git clone https://github.com/dmlc/mshadow', shell = True)
-        subprocess.call('cd ..; cp make/readthedocs.mk config.mk', shell = True)
-        subprocess.call('cd ..; rm -rf build', shell = True)
+        if READTHEDOCS_BUILD:
+            subprocess.call('cd ..; cp make/readthedocs.mk config.mk', shell = True)
+            subprocess.call('cd ..; rm -rf build', shell = True)
         retcode = subprocess.call("cd %s; make" % folder, shell = True)
         if retcode < 0:
             sys.stderr.write("build terminated by signal %s" % (-retcode))
     except OSError as e:
         sys.stderr.write("build execution failed: %s" % e)
 
-if os.environ.get('READTHEDOCS', None) == 'True':
+if READTHEDOCS_BUILD or not os.path.exists('../recommonmark'):
     subprocess.call('cd ..; rm -rf recommonmark;' +
                     'git clone https://github.com/tqchen/recommonmark', shell = True)
